Refactor MigrationHelper to delegate logic to helper classes
Streamlined the MigrationHelper class by delegating migration logic to specialized helper classes (CaseMigrationHelper, TaskMigrationHelper, and PetriNetMigrationHelper). This improves code modularity, maintainability, and reduces redundancy by centralizing shared operations.
